Add files via upload
This Dashboard Displays information of historical wildfires in Australia's seven regions between 2005 and 2020.
The dashboard provides the option to choose one of the seven regions and then a year. The dashboard then responds with a pie chart for the estimated fire area and a bar graph of the average pixels for presumed vegetation for the year and region selected.